requirements.txt
moviepy==1.0.3
PyMuPDF
pysrt
Pillow==9.5.0